Delete re-definition of lisp-indent-function for something simpler
The compiler kept complaining about how the variable
‘calculate-lisp-indent-last-sexp’ was “free,” which is due to (according to [1]) the
fact that it’s dynamically bound, as opposed to lexical – so it’s actually defined
further up the function call-chain, meaning it works but the compiler complains
about it.  I didn’t like all the complaining, so I tried installing ‘el-patch’ per
this StackOverflow discussion[2], thinking that a package built allegedly
specifically for this purpose would fix the dynamic/lexical binding mess, but of
course it didn’t.  So I just went and applied the changes in abo-abo’s answer[3],
which set the indent-functions for elisp to be the same as cl.  C’est la vie.

[1]: https://emacs.stackexchange.com/questions/52782/
[2]: https://stackoverflow.com/questions/22166895/
[3]: https://stackoverflow.com/a/22167050