Add Forge
Forge stores its information in an SQLite database, and on Windows,
that's not just available.  I found a solution in Win Builds
(http://win-builds.org/doku.php), though I think I could've just
downloaded a gcc binary and added that to my PATH, or something.  It
probably requires further research.  Anyway, that's why the exec-path
is updated as well.